3_busVoiceProject/requirements.txt
# Streamlit Cloud 배포용 최소 요구사항
streamlit==1.46.1
requests==2.32.4
python-dotenv==1.1.1

# LangChain 및 RAG 관련
langchain==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langchain-community==0.3.26

# ChromaDB (CPU 버전)
chromadb==1.0.13

# 기본 데이터 처리
pandas==2.3.0
numpy==2.3.1
